+++ CMakeLists.txt
@@ -6,7 +6,6 @@ set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
 
 set(EXECUTABLE_NAME "ZX_RGBI_TO_VGA_HDMI")
-set(FW_VERSION "v1.4.3-PICO-SDK")
 set(PICO_BOARD pico CACHE STRING "Board type")
 
 # Initialize pico_sdk from installed location
@@ -36,7 +35,7 @@ project(ZX_RGBI_TO_VGA_HDMI C CXX ASM)
 pico_sdk_init()
 
 # Add executable
-add_executable(${EXECUTABLE_NAME} main.cpp vga.c dvi.c g_config.c rgb_capture.c v_buf.c serial.cpp gpio.cpp)
+add_executable(${EXECUTABLE_NAME} dvi.c g_config.c gpio.cpp main.cpp rgb_capture.c serial_menu.cpp Serial.cpp settings.c v_buf.c vga.c video_output.c)
 
 # Generate PIO header
 pico_generate_pio_header(${EXECUTABLE_NAME} ${CMAKE_CURRENT_LIST_DIR}/programs.pio)
@@ -49,11 +48,13 @@ pico_enable_stdio_usb(${EXECUTABLE_NAME} 1)
 target_include_directories(${EXECUTABLE_NAME} PRIVATE ${CMAKE_CURRENT_LIST_DIR})
 
 # Add any user requested libraries
-target_link_libraries(${EXECUTABLE_NAME} hardware_pio hardware_dma hardware_flash pico_multicore pico_unique_id pico_stdlib pico_bootsel_via_double_reset pico_usb_reset_interface)
+target_link_libraries(${EXECUTABLE_NAME} hardware_pio hardware_dma hardware_flash pico_multicore pico_unique_id  pico_stdlib)
+# pico_bootsel_via_double_reset 
+# pico_usb_reset_interface     
 
-target_compile_options(${EXECUTABLE_NAME} PRIVATE -O3 -DFW_VERSION=\"${FW_VERSION}\")
+target_compile_options(${EXECUTABLE_NAME} PRIVATE -O3 )
 
 # Add linker flag to print memory usage
-target_link_options(${EXECUTABLE_NAME} PRIVATE -Wl,--print-memory-usage)
+target_link_options(${EXECUTABLE_NAME} PRIVATE -Wl,--print-memory-usage )
 
 pico_add_extra_outputs(${EXECUTABLE_NAME})
