+++ README.md
@@ -5,8 +5,8 @@ This repository contains code from [zx-rgbi-to-vga-hdmi](https://github.com/osem
 
 For detailed hardware and original software information, see the upstream projects:  
 
-- [RGBI_TO_VGA](https://github.com/tchv71/RGBI_TO_VGA)
-- [ZX_RGBI2VGA-HDMI](https://github.com/AlexEkb4ever/ZX_RGBI2VGA-HDMI/)
+🔗 [RGBI_TO_VGA](https://github.com/tchv71/RGBI_TO_VGA)
+🔗 [ZX_RGBI2VGA-HDMI](https://github.com/AlexEkb4ever/ZX_RGBI2VGA-HDMI/)
 
 ---
 
@@ -15,16 +15,22 @@ For detailed hardware and original software information, see the upstream projec
 ### Software
 
 - **Video Output:**
-  - VGA output with selectable resolutions: 640x480, 800x600, 1024x768, and 1280x1024.
-  - HDMI (DVI) output support at a fixed resolution of 640x480.
-  - Optional scanline effect at higher resolutions for a retro look.
+  - VGA output with selectable resolutions: 640×480 @60Hz, 800×600 @60Hz, 1024×768 @60Hz, 1280×1024 @60Hz.
+  - HDMI (DVI) resolutions: 640×480 @60Hz and 720×576 @50Hz.
+  - Optional scanline effect on the VGA output at higher resolutions for a retro look.
   - "NO SIGNAL" message when no input is detected.
 - **Configuration via Serial Terminal:**
   - Text-based menu system.
   - Frequency presets for self-synchronizing capture mode (supports ZX Spectrum 48K/128K pixel clocks).
-  - Real-time adjustment of all parameters (changes are applied immediately).
-  - Settings can be saved to flash memory without a restart.
+  - Real-time adjustment of all parameters (changes applied immediately).
+  - Settings can be saved to flash memory without restart.
 - **Test/Welcome Screen:** Styled after the ZX Spectrum 128K.
+
+### Hardware
+
+- **Analog to Digital Conversion:** Converts analog RGB to digital RGBI.
+  - Based on the project:  
+🔗 [RGBtoHDMI](https://github.com/hoglet67/RGBtoHDMI)
 
 ---
