+++ dvi.h
@@ -1,3 +1,4 @@
 #pragma once
 
 void start_dvi(video_mode_t);
+void stop_dvi();
