+++ rgb_capture.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #define CAP_LINE_LENGTH 1024
+// the number of DMA buffers can be increased if there is image fluttering
 #define CAP_DMA_BUF_CNT 8
 #define CAP_DMA_BUF_SIZE (CAP_LINE_LENGTH * CAP_DMA_BUF_CNT)
 
@@ -13,6 +14,5 @@ int16_t set_capture_shY(int16_t);
 int8_t set_capture_delay(int8_t);
 void set_pin_inversion_mask(uint8_t);
 void set_video_sync_mode(bool);
-void check_settings(settings_t *);
-void start_capture(settings_t *);
+void start_capture();
 void stop_capture();
