+++ settings.h
@@ -1,0 +1,5 @@
+#pragma once
+
+void check_settings(settings_t *);
+void load_settings(settings_t *);
+void save_settings(settings_t *);
